requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Security and JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Redis
redis==5.0.1

# HTTP client
httpx==0.25.2

# Telegram Bot
python-telegram-bot==20.7

# System monitoring
psutil==5.9.6

# Environment variables
python-dotenv==1.0.0

# Date utilities
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Linting (optional)
flake8==6.1.0
black==23.12.0
